environment variable implementation
