+++ templates/test.html
@@ -1,0 +1,187 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<title>PropertyProsperity</title>
+
+<link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
+
+</head>
+<body>
+
+<header class='navbar'>
+  <h2 class='navbar-heading'>PropertyBitch</h2>
+  <div class='navbar-buttons'>
+    <button style="padding-right: 20px;">Home</button>
+    <button style="padding-right: 20px; padding-left: 20px;">About</button>
+    <button style="padding-right: 20px; padding-left: 20px;">Contact</button>
+    <button style="padding-right: 20px; padding-left: 20px;">Buy</button>
+  </div>
+</header>
+
+<section class="w-full py-12 md:py-24 xl:py-32">
+  
+    <div id="stuff" class="container px-4 md:px-6 flex flex-col items-center space-y-4 text-center">
+    <div id="stuff" class="">
+      <h1 id="head" class="text-3xl font-bold tracking-tighter sm:text-5xl xl:text-6xl/none">Find your dream home</h1>
+      <p id="sub" class="mx-auto max-w-[600px] text-gray-500 dark:text-gray-400">
+        Discover the perfect place to call your own<br>
+        with our trusted listings.
+      </p>
+    </div>
+    <div class="flex flex-col gap-2 md:flex-row">
+      
+      <form action="/data" method="post">
+        <input type="text" name="input" placeholder="bc">
+        <button type="Submit" onclick="update_nut()" class="btn btn-primary btn-lg">Submit2</button>
+      </form>
+<br>
+      <form action="/aimset" method="post">
+        <input type="text" name="Calories" placeholder="bc">
+        <input type="text" name="Protein" placeholder="mc">
+        <input type="text" name="Fats" placeholder="bc">
+        <input type="text" name="Sugar" placeholder="mc">
+        <input type="text" name="Fiber" placeholder="bc">
+        <input type="text" name="Carbs" placeholder="mc">
+        <button type="Submit" onclick="aimset()" class="btn btn-primary btn-lg">Goal set</button>
+      </form>
+
+      <form action="/reset" method="post">
+        <button onclick="reset()" class="btn btn-primary btn-lg">reset</button>
+      </form>
+<br>
+<br>
+      <form action="/generate" method="post">
+        <button type="Submit" onclick="generate()" class="btn btn-primary btn-lg">generate</button>
+      </form>
+
+
+    <p id="sub">{{content}}</p>
+
+       
+
+      </div>
+    </div>
+  </div> 
+</section>
+
+<script>
+
+    function update_nut() {
+      fetch('/data')
+          .then(response => response.text())
+          .then(data => console.log(data))
+          .catch(error => console.error('Error:', error));
+      
+        location.reload();
+  }
+
+  function aimset() {
+    fetch('/aimset')
+        .then(response => response.text())
+        .then(data => console.log(data))
+        .catch(error => console.error('Error:', error));
+    
+      location.reload();
+}
+
+  function reset() {
+    fetch('/reset')
+        .then(response => {
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+            return response.text(); // Assuming the response is text
+        })
+        .then(data => {
+            // Handle the response data here
+            console.log('Reset successful');
+            // You might want to update the UI or take other actions based on the response
+        })
+        .catch(error => {
+            // Handle errors here
+            console.error('There was a problem with the fetch operation:', error);
+        });
+      }
+
+  function generate() {
+    fetch('/generate')
+        .then(response => {
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+            return response.text(); // Assuming the response is text
+        })
+        .then(data => {
+            // Handle the response data here
+            console.log('Reset successful');
+            // You might want to update the UI or take other actions based on the response
+        })
+        .catch(error => {
+            // Handle errors here
+            console.error('There was a problem with the fetch operation:', error);
+        });
+}
+
+</script>
+
+
+<style>
+  body {
+    margin: 0;
+    font-family: "karla", sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    width: 100%;
+    background-color: navy;
+    background-size: cover;
+  }
+
+  .navbar {
+    display: flex;
+    align-items: center;
+    height: 15%;
+    color: black;
+    box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.25);
+    background-color: white;
+  }
+
+  .navbar-heading {
+    padding-left: 20px;
+  }
+
+  .navbar-buttons {
+    margin-left: auto;
+    padding-right: 20px;
+  }
+
+  #head {
+    margin-top: 10%;
+    display: flex;
+    font-size: 400%;
+    justify-content: center;
+    color: white;
+  }
+
+  #sub {
+    font-size: 180%;
+    display: flex;
+    justify-content: center;
+    color: white;
+  }
+
+  #subby {
+    font-size: 120%;
+    display: flex;
+    justify-content: center;
+    color: white;
+  }
+
+  #butt {
+    width: 50%;
+  }
+</style>
+
+</body>
+</html>
